v1.0.0/backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
sqlalchemy>=2.0.30,<3.0.0
pydantic>=2.7.1,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
opencv-python-headless>=4.10.0,<5.0.0
ultralytics>=8.2.0
groq>=0.9.0
google-generativeai>=0.7.2
requests>=2.31.0,<3.0.0
pillow>=11.0.0
websockets>=12.0,<16.0
python-multipart>=0.0.9,<1.0.0
aiofiles>=23.2.1,<25.0.0
httpx>=0.27.0,<1.0.0
numpy>=2.0.0
